requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
altgraph==0.17.2
anaconda==0.0.1.1
ansicon==1.89.0
arrow==1.2.0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
backtrader==1.9.76.123
blessed==1.17.6
botocore==1.21.53
bravado==10.4.1
bravado-core==5.17.0
Bybit==0.2.12
cachetools==4.2.4
ccxt==1.57.88
certifi==2021.10.8
cffi==1.14.6
chalice==1.26.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
distlib==0.3.3
filelock==3.3.0
fqdn==1.5.1
future==0.18.2
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.20.1
idna==3.2
importlib-metadata==4.8.1
inquirer==2.7.0
isoduration==20.11.0
jinxed==1.1.0
jmespath==0.10.0
jsonpointer==2.1
jsonref==0.2
jsonschema==3.1.1
kiwisolver==1.3.2
matplotlib==3.4.3
monotonic==1.6
msgpack==1.0.2
msgpack-python==0.5.6
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pefile==2021.9.3
Pillow==8.3.2
platformdirs==2.4.0
protobuf==3.18.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readchar==2.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rsa==4.7.2
schedule==1.1.0
simplejson==3.17.5
six==1.16.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
ta==0.7.0
typing-extensions==3.10.0.2
uri-template==1.1.0
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==20.8.1
wcwidth==0.2.5
webcolors==1.11.1
yarl==1.6.3
zipp==3.6.0